ponderada2/requirements.txt
Flask==2.2.5
psycopg2_binary==2.9.8
